requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
qdrant-client==1.9.2
openai==1.3.5
PyPDF2==3.0.1
python-multipart==0.0.6
requests==2.31.0
numpy==1.26.3
httpx==0.25.2
<<<<<<< HEAD
pydantic==2.5.2
=======
pydantic==2.5.2
python-multipart==0.0.6
>>>>>>> 002-ai-book-rag-chatbot
